refactor: extract handleSystemRunPrepare to fix type_body_length lint
